feat: change icons and transform copy to clipboard logic into bytemd plugin
